Spellslots implemented better
Made changes so spellslots are refreshed without refreshing the character tabs.
Changed combatant to have he description available. This will show the whole transformation stack for a friendly character, but only the description for the last in the transformatin stack for the others.
Changed the description shown in slave panels for unfiendly characters to provide them with de notification that there is no more information available.