Changes for better start
At the start files are now loaded in multiple threads. This shortens the startup time dramatically for environments with multiple cores.
Also removed the last frame that could be a panel.
Added the option to add another combatant to a battle.
Added speed to information panel for all characters (battle).