Hidden combatants
Added the option for combatants to be hidden. In the DM screen these are shown as regular combatants. On the player screen these are hidden unless thay are friendly. The first combatant on the player screen also only has a blue border if there are no hidden combatants before it.